+++ src/output.cpp
@@ -7,7 +7,7 @@
 
 
 void print_text (struct Text_t text) {
-    for (unsigned i = 0; i < text.num_lines; i++) {
+    for (size_t i = 0; i < text.num_lines; i++) {
         char str_no[10] = ""; // 10 - maximum length of unsigned int in base 10
         itoa (i, str_no, 10);
 
@@ -21,14 +21,14 @@ void print_str (char str_name[], const char *string) {
     assert (string);
 
     printf ("%s = ", str_name);
-    for (unsigned i = 0; i < strlen(string); i++)
+    for (size_t i = 0; i < strlen(string); i++)
         printf ("%c(%3d) ", string[i], string[i]);
     printf ("\n");
 }
 
-void print_pointer_array (char **ptr_array, unsigned n_lines) {
+void print_pointer_array (char **ptr_array, size_t n_lines) {
     printf ("\nArray of pointers:\n");
-    for (unsigned i = 0; i < n_lines; i++) {
+    for (size_t i = 0; i < n_lines; i++) {
         assert (i < n_lines);
 
         printf ("Str %d = %lu\n", i, (unsigned long) ptr_array[i]);
