+++ src/sorting.cpp
@@ -3,8 +3,8 @@
 
 
 struct Text_t bubble_sort (struct Text_t sorting_text) {
-    for (unsigned i = sorting_text.num_lines; i > 0; i--) {
-        for (unsigned j = 0; j < i - 1; j++) {
+    for (size_t i = sorting_text.num_lines; i > 0; i--) {
+        for (size_t j = 0; j < i - 1; j++) {
             assert (j + 1 < sorting_text.num_lines);
 
             if (my_strcmp (sorting_text.ptr_array[j], sorting_text.ptr_array[j + 1]) > 0) {
